chore: move Arcjet to root layout
