+++ src/app/[locale]/(auth)/layout.tsx
@@ -1,20 +1,7 @@
-import arcjet, { detectBot, request } from '@/libs/Arcjet';
 import { routing } from '@/libs/i18nNavigation';
 import { enUS, frFR } from '@clerk/localizations';
 import { ClerkProvider } from '@clerk/nextjs';
 import { setRequestLocale } from 'next-intl/server';
-
-const aj = arcjet.withRule(
-  detectBot({
-    mode: 'LIVE',
-    // Block all bots except the following
-    allow: [
-      // See https://docs.arcjet.com/bot-protection/identifying-bots
-      'CATEGORY:PREVIEW', // Allow preview links to show OG images
-      'CATEGORY:MONITOR', // Allow uptime monitoring services
-    ],
-  }),
-);
 
 export default async function AuthLayout(props: {
   children: React.ReactNode;
@@ -37,19 +24,6 @@ export default async function AuthLayout(props: {
     signUpUrl = `/${locale}${signUpUrl}`;
     dashboardUrl = `/${locale}${dashboardUrl}`;
     afterSignOutUrl = `/${locale}${afterSignOutUrl}`;
-  }
-
-  const req = await request();
-  const decision = await aj.protect(req);
-
-  // These errors are handled by the global error boundary, but you could also
-  // redirect or show a custom error page
-  if (decision.isDenied()) {
-    if (decision.reason.isBot()) {
-      throw new Error('No bots allowed');
-    }
-
-    throw new Error('Access denied');
   }
 
   return (
