+++ src/app/[locale]/(marketing)/layout.tsx
@@ -1,22 +1,8 @@
 import { DemoBanner } from '@/components/DemoBanner';
 import { LocaleSwitcher } from '@/components/LocaleSwitcher';
-import arcjet, { detectBot, request } from '@/libs/Arcjet';
 import { BaseTemplate } from '@/templates/BaseTemplate';
 import Link from 'next/link';
 import { getTranslations, setRequestLocale } from 'next-intl/server';
-
-const aj = arcjet.withRule(
-  detectBot({
-    mode: 'LIVE',
-    // Block all bots except the following
-    allow: [
-      // See https://docs.arcjet.com/bot-protection/identifying-bots
-      'CATEGORY:SEARCH_ENGINE', // Allow search engines
-      'CATEGORY:PREVIEW', // Allow preview links to show OG images
-      'CATEGORY:MONITOR', // Allow uptime monitoring services
-    ],
-  }),
-);
 
 export default async function Layout(props: {
   children: React.ReactNode;
@@ -28,19 +14,6 @@ export default async function Layout(props: {
     locale,
     namespace: 'RootLayout',
   });
-
-  const req = await request();
-  const decision = await aj.protect(req);
-
-  // These errors are handled by the global error boundary, but you could also
-  // redirect or show a custom error page
-  if (decision.isDenied()) {
-    if (decision.reason.isBot()) {
-      throw new Error('No bots allowed');
-    }
-
-    throw new Error('Access denied');
-  }
 
   return (
     <>
