+++ src/app/[locale]/layout.tsx
@@ -1,5 +1,6 @@
 import type { Metadata } from 'next';
 import { DemoBadge } from '@/components/DemoBadge';
+import arcjet, { detectBot, request } from '@/libs/Arcjet';
 import { routing } from '@/libs/i18nNavigation';
 import { notFound } from 'next/navigation';
 import { NextIntlClientProvider } from 'next-intl';
@@ -35,6 +36,19 @@ export function generateStaticParams() {
   return routing.locales.map(locale => ({ locale }));
 }
 
+const aj = arcjet.withRule(
+  detectBot({
+    mode: 'LIVE',
+    // Block all bots except the following
+    allow: [
+      // See https://docs.arcjet.com/bot-protection/identifying-bots
+      'CATEGORY:SEARCH_ENGINE', // Allow search engines
+      'CATEGORY:PREVIEW', // Allow preview links to show OG images
+      'CATEGORY:MONITOR', // Allow uptime monitoring services
+    ],
+  }),
+);
+
 export default async function RootLayout(props: {
   children: React.ReactNode;
   params: Promise<{ locale: string }>;
@@ -46,6 +60,19 @@ export default async function RootLayout(props: {
   }
 
   setRequestLocale(locale);
+
+  const req = await request();
+  const decision = await aj.protect(req);
+
+  // These errors are handled by the global error boundary, but you could also
+  // redirect or show a custom error page
+  if (decision.isDenied()) {
+    if (decision.reason.isBot()) {
+      throw new Error('No bots allowed');
+    }
+
+    throw new Error('Access denied');
+  }
 
   // Using internationalization in Client Components
   const messages = await getMessages();
